refactor: remove unused ESLint and Prettier configuration files and update exports
ref #30
