--- src/eslint/configs/index.ts
@@ -1,4 +1,0 @@
-import { base } from './base';
-import { djblackeagle } from './djblackeagle';
-
-export const configs = { base, djblackeagle };
